requirements.txt
fastapi
uvicorn
jinja2
pillow
torch
torchvision
python-multipart
